Moved Azure Storage code that was only being used by the integration tests into the test project to clean up the client.
